# Conflicts:
#	src/main/java/ru/skypro/homework/controller/AdController.java

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -3,11 +3,9 @@ package ru.skypro.homework.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.Ad;
 import ru.skypro.homework.dto.Ads;
@@ -15,12 +13,13 @@ import ru.skypro.homework.dto.CreateOrUpdateAd;
 import ru.skypro.homework.dto.ExtendedAd;
 import ru.skypro.homework.entity.AdEntity;
 import ru.skypro.homework.entity.UserEntity;
+import ru.skypro.homework.exception.ForbiddenException;
 import ru.skypro.homework.mapper.AdMapper;
 import ru.skypro.homework.repository.AdRepository;
 import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.service.AdService;
 
-import javax.transaction.Transactional;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -74,24 +73,34 @@ public class AdServiceImpl implements AdService {
         return extendedAd;
     }
 
-    public Ad updateAd(Integer id, CreateOrUpdateAd adDetails){
+    public Ad updateAd(Integer id, CreateOrUpdateAd adDetails, UserDetails userDetails){
         AdEntity adEntity = adRepository.findById(id).orElseThrow();
-        AdEntity adEntityUpdated = adMapper.createOrUpdateAdDTOToAd(adDetails);
-        adEntity.setTitle(adEntityUpdated.getTitle());
-        adEntity.setPrice(adEntityUpdated.getPrice());
-        adEntity.setDescription(adEntityUpdated.getDescription());
+        if(checkAccess(userDetails, adEntity)) {
+            AdEntity adEntityUpdated = adMapper.createOrUpdateAdDTOToAd(adDetails);
+            adEntity.setTitle(adEntityUpdated.getTitle());
+            adEntity.setPrice(adEntityUpdated.getPrice());
+            adEntity.setDescription(adEntityUpdated.getDescription());
 
-        adRepository.save(adEntity);
+            adRepository.save(adEntity);
 
-        return adMapper.adToAdDTO(adEntity);
+            return adMapper.adToAdDTO(adEntity);
+        } else throw new ForbiddenException();
     }
 
-    public ResponseEntity<String> removeAd(Integer id){
+    public ResponseEntity<String> removeAd(Integer id, UserDetails userDetails) {
         Optional<AdEntity> adEntityOptional = adRepository.findById(id);
-        if(adEntityOptional.isPresent()){
-            adRepository.deleteById(id);
-            return new ResponseEntity<>("Коментарий удален", HttpStatus.OK);
-        } else return new ResponseEntity<>("Коментарий не найден", HttpStatus.NOT_FOUND);
+       if (checkAccess(userDetails, adEntityOptional.get())) {
+            if (adEntityOptional.isPresent()) {
+                adRepository.deleteById(id);
+                return new ResponseEntity<>("Коментарий удален", HttpStatus.OK);
+            } else return new ResponseEntity<>("Коментарий не найден", HttpStatus.NOT_FOUND);
+       } else throw new ForbiddenException();
+    }
+
+    private boolean checkAccess(UserDetails userDetails, AdEntity adEntity) {
+        return userDetails.getUsername().equals(adEntity.getAuthor().getUsername())
+               || userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"));
+
     }
 
 }
